+++ controllers/routes/api/index.js
@@ -3,6 +3,7 @@ const userRoutes = require("./userRoutes");
 const studentRoutes = require("./studentRoutes");
 const instructorRoutes = require("./instructorRoutes");
 const courseRoutes = require("./courseRoutes");
+const tagRoutes = require("./tagRoutes");
 
 // Route : api/
 
@@ -10,5 +11,6 @@ router.use("/users", userRoutes);
 router.use("/students", studentRoutes);
 router.use("/instructors", instructorRoutes);
 router.use("/courses", courseRoutes);
+router.use("/tags", tagRoutes);
 
 module.exports = router;
