+++ controllers/services/courseServices.js
@@ -1,5 +1,6 @@
 const { Op } = require("sequelize");
 const models = require("../../models");
+const courseTagServices = require("./courseTagServices");
 const sanitize = require("./sanitize");
 const { filterObjectByKeys } = require("../utils/object");
 const acceptedKeys = ["course_name", "active", "instructor_id"];
@@ -151,77 +152,11 @@ const removeStudent = async (student, courseId) => {
 //    - active
 //    - tags: Array<tag_id> - assumed to be all the tags that this course has
 //  - courseId
-
-const getCourseTags = async (courseId) => {
-  const result = await models.CourseTag.findAll({
-    where: { course_id: courseId },
-  });
-  return result;
-};
-
-const isIn = (tagId, courseTags) => {
-  const tagIds = courseTags.map(({ tag_id }) => tag_id);
-  return tagIds.includes(tagId);
-};
-
-const getCourseTagIdsToRemove = (oldCourseTags, newCourseTags) => {
-  const toRemove = oldCourseTags.filter(({ tag_id }) => {
-    const entry = isIn(tag_id, newCourseTags);
-    return entry ? false : true; // remove one not in tags
-  });
-  const result = toRemove.map(({ id }) => id);
-  return result;
-};
-
-// remove old course tags not in the new one
-const removeCourseTags = async (oldCourseTags, newCourseTags) => {
-  const idsToRemove = getCourseTagIdsToRemove(oldCourseTags, newCourseTags);
-  const result = await models.CourseTag.destroy({ where: { id: idsToRemove } });
-  return result;
-};
-
-// create new course tags data
-const createNewCourseTags = (tags, courseId) => {
-  const result = tags.map((tag) => {
-    return { tag_id: tag, course_id: courseId };
-  });
-  return result;
-};
-
-const getCourseTagsToCreate = (oldCourseTags, newCourseTags) => {
-  const result = newCourseTags.filter(({ tag_id }) => {
-    const entry = isIn(tag_id, oldCourseTags);
-    return entry ? false : true; // add one not in the old ones
-  });
-  return result;
-};
-
-// add new course tags not in the old one
-const addCourseTags = async (oldCourseTags, newCourseTags) => {
-  const toCreate = getCourseTagsToCreate(oldCourseTags, newCourseTags);
-  const result = await models.CourseTag.bulkCreate(toCreate);
-  return result;
-};
-
-// update course tag
-const updateCourseTags = async (tags, courseId) => {
-  const oldCourseTags = await getCourseTags(courseId);
-  const newCourseTags = createNewCourseTags(tags, courseId);
-
-  console.log("\ncourse services: ", newCourseTags);
-
-  await removeCourseTags(oldCourseTags, newCourseTags);
-  await addCourseTags(oldCourseTags, newCourseTags);
-};
-
-// combine everything together
 const update = async (updateData, courseId) => {
   await models.Course.update(updateData, { where: { id: courseId } });
 
   if ("tags" in updateData && updateData.tags.length) {
-    console.log("\ncourse services:  update course tags\n");
-
-    await updateCourseTags(updateData.tags, courseId);
+    await courseTagServices.update(updateData.tags, courseId);
   }
   const result = await getOne(courseId);
   return result;
